downgrade go-difflib and go-spew to tagged releases
These dependencies were updated to "master" in some modules we depend on,
but have no code-changes since their last release. Unfortunately, this also
causes a ripple effect, forcing all users of the containerd module to also
update these dependencies to an unrelease / un-tagged version.

Both these dependencies will unlikely do a new release in the near future,
so exclude these versions so that we can downgrade to the current release.

For additional details, see [this PR][1] and links mentioned in it.

[1]: https://github.com/kubernetes-sigs/kustomize/pull/5830#issuecomment-2569960859

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>